fix(senpi-codemode): split Python kernel contracts
